+++ Userland/Libraries/LibGUI/GML/SyntaxHighlighter.cpp
@@ -47,11 +47,11 @@ void SyntaxHighlighter::rehighlight(Palette const& palette)
 
     Vector<Token> folding_region_start_tokens;
 
-    Vector<GUI::TextDocumentSpan> spans;
-    Vector<GUI::TextDocumentFoldingRegion> folding_regions;
+    Vector<Syntax::TextDocumentSpan> spans;
+    Vector<Syntax::TextDocumentFoldingRegion> folding_regions;
 
     for (auto& token : tokens) {
-        GUI::TextDocumentSpan span;
+        Syntax::TextDocumentSpan span;
         span.range.set_start({ token.m_start.line, token.m_start.column });
         span.range.set_end({ token.m_end.line, token.m_end.column });
         span.attributes = style_for_token_type(palette, token.m_type);
@@ -65,7 +65,7 @@ void SyntaxHighlighter::rehighlight(Palette const& palette)
         } else if (token.m_type == Token::Type::RightCurly) {
             if (!folding_region_start_tokens.is_empty()) {
                 auto left_curly = folding_region_start_tokens.take_last();
-                GUI::TextDocumentFoldingRegion region;
+                Syntax::TextDocumentFoldingRegion region;
                 region.range.set_start({ left_curly.m_end.line, left_curly.m_end.column });
                 region.range.set_end({ token.m_start.line, token.m_start.column });
                 folding_regions.append(region);
