+++ Userland/Libraries/LibSyntax/CMakeLists.txt
@@ -1,4 +1,5 @@
 set(SOURCES
+    Document.cpp
     Highlighter.cpp
     Language.cpp
 )
